# internal/api/http/server/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package server

import (
	"fmt"
)

func (s *ErrorResponseStatusCode) Error() string {
	return fmt.Sprintf("code %d: %+v", s.StatusCode, s.Response)
}

// Ref: #/AuthLoginRequestSchema
type AuthLoginRequestSchema struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

// GetEmail returns the value of Email.
func (s *AuthLoginRequestSchema) GetEmail() string {
	return s.Email
}

// GetPassword returns the value of Password.
func (s *AuthLoginRequestSchema) GetPassword() string {
	return s.Password
}

// SetEmail sets the value of Email.
func (s *AuthLoginRequestSchema) SetEmail(val string) {
	s.Email = val
}

// SetPassword sets the value of Password.
func (s *AuthLoginRequestSchema) SetPassword(val string) {
	s.Password = val
}

type BearerAuth struct {
	Token string
}

// GetToken returns the value of Token.
func (s *BearerAuth) GetToken() string {
	return s.Token
}

// SetToken sets the value of Token.
func (s *BearerAuth) SetToken(val string) {
	s.Token = val
}

// Ref: #/components/schemas/ErrorResponse
type ErrorResponse struct {
	// Текст ошибки.
	ErrorMessage string `json:"error_message"`
}

// GetErrorMessage returns the value of ErrorMessage.
func (s *ErrorResponse) GetErrorMessage() string {
	return s.ErrorMessage
}

// SetErrorMessage sets the value of ErrorMessage.
func (s *ErrorResponse) SetErrorMessage(val string) {
	s.ErrorMessage = val
}

// ErrorResponseStatusCode wraps ErrorResponse with StatusCode.
type ErrorResponseStatusCode struct {
	StatusCode int
	Response   ErrorResponse
}

// GetStatusCode returns the value of StatusCode.
func (s *ErrorResponseStatusCode) GetStatusCode() int {
	return s.StatusCode
}

// GetResponse returns the value of Response.
func (s *ErrorResponseStatusCode) GetResponse() ErrorResponse {
	return s.Response
}

// SetStatusCode sets the value of StatusCode.
func (s *ErrorResponseStatusCode) SetStatusCode(val int) {
	s.StatusCode = val
}

// SetResponse sets the value of Response.
func (s *ErrorResponseStatusCode) SetResponse(val ErrorResponse) {
	s.Response = val
}

// Merged schema.
// Ref: #/Item
type Item struct {
	Title       string    `json:"title"`
	Description OptString `json:"description"`
	ListID      int64     `json:"list_id"`
	ID          int64     `json:"id"`
}

// GetTitle returns the value of Title.
func (s *Item) GetTitle() string {
	return s.Title
}

// GetDescription returns the value of Description.
func (s *Item) GetDescription() OptString {
	return s.Description
}

// GetListID returns the value of ListID.
func (s *Item) GetListID() int64 {
	return s.ListID
}

// GetID returns the value of ID.
func (s *Item) GetID() int64 {
	return s.ID
}

// SetTitle sets the value of Title.
func (s *Item) SetTitle(val string) {
	s.Title = val
}

// SetDescription sets the value of Description.
func (s *Item) SetDescription(val OptString) {
	s.Description = val
}

// SetListID sets the value of ListID.
func (s *Item) SetListID(val int64) {
	s.ListID = val
}

// SetID sets the value of ID.
func (s *Item) SetID(val int64) {
	s.ID = val
}

// Ref: #/ItemCommonData
type ItemCommonData struct {
	Title       string    `json:"title"`
	Description OptString `json:"description"`
}

// GetTitle returns the value of Title.
func (s *ItemCommonData) GetTitle() string {
	return s.Title
}

// GetDescription returns the value of Description.
func (s *ItemCommonData) GetDescription() OptString {
	return s.Description
}

// SetTitle sets the value of Title.
func (s *ItemCommonData) SetTitle(val string) {
	s.Title = val
}

// SetDescription sets the value of Description.
func (s *ItemCommonData) SetDescription(val OptString) {
	s.Description = val
}

// Merged schema.
// Ref: #/ItemCreateInput
type ItemCreateInput struct {
	Title       string    `json:"title"`
	Description OptString `json:"description"`
	ListID      int64     `json:"list_id"`
}

// GetTitle returns the value of Title.
func (s *ItemCreateInput) GetTitle() string {
	return s.Title
}

// GetDescription returns the value of Description.
func (s *ItemCreateInput) GetDescription() OptString {
	return s.Description
}

// GetListID returns the value of ListID.
func (s *ItemCreateInput) GetListID() int64 {
	return s.ListID
}

// SetTitle sets the value of Title.
func (s *ItemCreateInput) SetTitle(val string) {
	s.Title = val
}

// SetDescription sets the value of Description.
func (s *ItemCreateInput) SetDescription(val OptString) {
	s.Description = val
}

// SetListID sets the value of ListID.
func (s *ItemCreateInput) SetListID(val int64) {
	s.ListID = val
}

// Merged schema.
// Ref: #/List
type List struct {
	ID    int64  `json:"id"`
	Name  string `json:"name"`
	Items []Item `json:"items"`
}

// GetID returns the value of ID.
func (s *List) GetID() int64 {
	return s.ID
}

// GetName returns the value of Name.
func (s *List) GetName() string {
	return s.Name
}

// GetItems returns the value of Items.
func (s *List) GetItems() []Item {
	return s.Items
}

// SetID sets the value of ID.
func (s *List) SetID(val int64) {
	s.ID = val
}

// SetName sets the value of Name.
func (s *List) SetName(val string) {
	s.Name = val
}

// SetItems sets the value of Items.
func (s *List) SetItems(val []Item) {
	s.Items = val
}

// Merged schema.
// Ref: #/ListCreateInput
type ListCreateInput struct {
	Items []ItemCommonData `json:"items"`
	Name  string           `json:"name"`
}

// GetItems returns the value of Items.
func (s *ListCreateInput) GetItems() []ItemCommonData {
	return s.Items
}

// GetName returns the value of Name.
func (s *ListCreateInput) GetName() string {
	return s.Name
}

// SetItems sets the value of Items.
func (s *ListCreateInput) SetItems(val []ItemCommonData) {
	s.Items = val
}

// SetName sets the value of Name.
func (s *ListCreateInput) SetName(val string) {
	s.Name = val
}

type ListListResponseSchema []List

// NewOptString returns new OptString with value set to v.
func NewOptString(v string) OptString {
	return OptString{
		Value: v,
		Set:   true,
	}
}

// OptString is optional string.
type OptString struct {
	Value string
	Set   bool
}

// IsSet returns true if OptString was set.
func (o OptString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptString) Reset() {
	var v string
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptString) SetTo(v string) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptString) Get() (v string, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// Ref: #/Token
type Token struct {
	AccessToken string `json:"access_token"`
}

// GetAccessToken returns the value of AccessToken.
func (s *Token) GetAccessToken() string {
	return s.AccessToken
}

// SetAccessToken sets the value of AccessToken.
func (s *Token) SetAccessToken(val string) {
	s.AccessToken = val
}